--- src/store/api-actions.ts
@@ -1,55 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-import { AxiosInstance } from 'axios';
-
-import { AppDispatch, State } from '../types/state';
-import { TOffer } from '../types/offer';
-import { APIRoute, AuthorizationStatus } from '../const';
-// import { loadOffers, requireAuthorization, setOffersDataLoadingStatus } from './action';
-import { TAuthData } from '../types/auth-data';
-import { TUserData } from '../types/user-data';
-import { saveToken } from '../services/token';
-
-const fetchOffers = createAsyncThunk<void, undefined, {
-  dispatch: AppDispatch;
-  state: State;
-  extra: AxiosInstance;
-}>(
-  'data/fetchOffers',
-  async (_arg, {dispatch, extra: api}) => {
-    dispatch(setOffersDataLoadingStatus(true));
-    const {data} = await api.get<TOffer[]>(APIRoute.Offers);
-    dispatch(loadOffers(data));
-    dispatch(setOffersDataLoadingStatus(false));
-  }
-);
-
-const fetchCheckAuthAction = createAsyncThunk<void, undefined, {
-  dispatch: AppDispatch;
-  state: State;
-  extra: AxiosInstance;
-}>(
-  'user/checkAuth',
-  async (_arg, {dispatch, extra: api}) => {
-    try {
-      await api.get(APIRoute.Login);
-      dispatch(requireAuthorization(AuthorizationStatus.Auth));
-    } catch {
-      dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
-    }
-  }
-);
-
-const fetchLoginAction = createAsyncThunk<void, TAuthData, {
-  dispatch: AppDispatch;
-  state: State;
-  extra: AxiosInstance;
-}>(
-  'user/checkAuth',
-  async ({email, password}, {dispatch, extra: api}) => {
-    const {data: {token}} = await api.post<TUserData>(APIRoute.Login, {email, password});
-    saveToken(token);
-    dispatch(requireAuthorization(AuthorizationStatus.Auth));
-  }
-);
-
-export {fetchOffers, fetchCheckAuthAction, fetchLoginAction};
