+++ src/store/offers/offers-selector.ts
@@ -1,4 +1,5 @@
 import { createSelector } from '@reduxjs/toolkit';
+
 import { State } from '../../types/state';
 import { NameSpace } from '../../const';
