+++ src/store/user/user-slice.ts
@@ -1,12 +1,12 @@
 import { createSlice } from '@reduxjs/toolkit';
 
 import { AuthorizationStatus, NameSpace } from '../../const';
-import { TAuthorizationStatus } from '../../types/authorization';
 import { fetchCheckAuth, fetchLogoutAuth, fetchPostLoginAction } from './user-action';
 import { dropToken, saveToken } from '../../services/token';
+import { TAuthStatus } from '../../types/authorization';
 
 type TInitialState = {
-  authorizationStatus: TAuthorizationStatus;
+  authorizationStatus: TAuthStatus;
   email: string | null;
 };
