--- src/types/auth-data.ts
@@ -1,4 +1,0 @@
-export type TAuthData = {
-  email: string;
-  password: string;
-}
