--- src/types/map-icon.ts
@@ -1,11 +1,0 @@
-export type TIconConfig = {
-  url: string;
-  width: number;
-  height: number;
-  anchorX: number;
-  anchorY: number;
-}
-
-export type TMapIcon = {
-  [K in 'Default' | 'Active']: TIconConfig;
-}
