--- src/types/map-location.ts
@@ -1,7 +1,0 @@
-type TMapLocation = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-};
-
-export type { TMapLocation };
