+++ src/types/state.ts
@@ -1,4 +1,8 @@
+import { LoadingStatus } from '../const';
 import { store } from '../store';
 
 export type State = ReturnType<typeof store.getState>;
+
 export type AppDispatch = typeof store.dispatch;
+
+export type TLoadingStatus = typeof LoadingStatus[keyof typeof LoadingStatus];
