+++ src/utils/map.ts
@@ -1,6 +1,6 @@
 import { Icon } from 'leaflet';
 
-import { TIconConfig } from '../types/map-icon';
+import { TIconConfig } from '../types/map';
 
 function createIcon(config: TIconConfig) {
   return new Icon({
